REFACTOR: Addition of StemNodes to capture SuffixExtensions (#8)
Currently, there was branch or leaf nodes (and null). The logic of suffix and extensions was all included in hash visitor.
This includes commiting to 1, stem, left values commitment and right values commitment.
However, the subcommitments will be useful in proofs.

Another reason to refactor is that path-extensions are only allowed at StemNodes, like in the current specifications.
Before, path-extensions could be done anywhere.

Another feature is support for StoredVerkleTrie. StoredNodes are now behaving correctly and can be retrieved recursively.

Finally, minor feature is that put returns the old value, which is helpful for trie-logs

Signed-off-by: Thomas Zamojski <thomas.zamojski@quadratic-labs.com>